`EuiTitle` Converted to TS
