--- src/components/title/index.js
@@ -1,3 +1,0 @@
-export {
-  EuiTitle,
-} from './title';
