+++ src/components/title/index.ts
@@ -1,0 +1,1 @@
+export { EuiTitle, EuiTitleSize } from './title';
